+++ src/test/java/com/vaadin/ui/textfield/demo/TextFieldView.java
@@ -13,12 +13,14 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.textfield;
+package com.vaadin.ui.textfield.demo;
 
 import com.vaadin.flow.demo.DemoView;
 import com.vaadin.router.Route;
 import com.vaadin.ui.common.HtmlImport;
 import com.vaadin.ui.html.Div;
+import com.vaadin.ui.textfield.GeneratedVaadinTextField;
+import com.vaadin.ui.textfield.TextField;
 
 /**
  * View for {@link GeneratedVaadinTextField} demo.
