+++ src/test/java/com/vaadin/ui/textfield/tests/PasswordFieldIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.textfield;
+package com.vaadin.ui.textfield.tests;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -22,6 +22,7 @@ import org.openqa.selenium.WebElement;
 
 import com.vaadin.flow.demo.ComponentDemoTest;
 import com.vaadin.testbench.By;
+import com.vaadin.ui.textfield.PasswordField;
 
 /**
  * Integration tests for the {@link PasswordField}.
