+++ src/test/java/com/vaadin/ui/textfield/tests/PasswordFieldValidationPage.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.textfield.it;
+package com.vaadin.ui.textfield.tests;
 
 import com.vaadin.flow.testutil.ValidationTestView;
 import com.vaadin.router.Route;
