+++ src/test/java/com/vaadin/ui/textfield/tests/TextAreaIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.textfield;
+package com.vaadin.ui.textfield.tests;
 
 import java.util.stream.IntStream;
 
@@ -24,6 +24,7 @@ import org.openqa.selenium.WebElement;
 
 import com.vaadin.flow.demo.ComponentDemoTest;
 import com.vaadin.testbench.By;
+import com.vaadin.ui.textfield.TextArea;
 
 /**
  * Integration tests for the {@link TextArea}.
